requirements.txt
Flask
python-pptx
google-generativeai
python-dotenv
gunicorn